Move websocket client headers to websocket instead of websocketBase
