+++ src/main/java/io/vertx/core/http/impl/WebSocketImplBase.java
@@ -64,7 +64,6 @@ public abstract class WebSocketImplBase<S extends WebSocketBase> implements WebS
   protected boolean closed;
   private Short closeStatusCode;
   private String closeReason;
-  private MultiMap headers;
 
   WebSocketImplBase(Http1xConnectionBase conn,
                     boolean supportsContinuation,
@@ -211,19 +210,6 @@ public abstract class WebSocketImplBase<S extends WebSocketBase> implements WebS
   public String closeReason() {
     synchronized (conn) {
       return closeReason;
-    }
-  }
-
-  @Override
-  public MultiMap headers() {
-    synchronized(conn) {
-      return headers;
-    }
-  }
-
-  void headers(MultiMap responseHeaders) {
-    synchronized(conn) {
-      this.headers = responseHeaders;
     }
   }
